backend update

+++ src/main/java/com/example/forums_backend/api/admin/AccountManageController.java
@@ -1,14 +1,11 @@
 package com.example.forums_backend.api.admin;
 
 import com.example.forums_backend.entity.Account;
-import com.example.forums_backend.entity.Post;
 import com.example.forums_backend.repository.AccountRepository;
 import com.example.forums_backend.service.AccountManagerService;
 import com.example.forums_backend.service.AccountService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,7 +14,7 @@ import static com.example.forums_backend.config.constant.route.AdminRoute.*;
 @RestController
 @RequestMapping(PREFIX_ADMIN_ROUTE)
 @RequiredArgsConstructor
-public class AccController {
+public class AccountManageController {
     @Autowired
     AccountManagerService accountManagerService;
 
@@ -27,23 +24,36 @@ public class AccController {
     @Autowired
     AccountService accountService;
 
+    @RequestMapping(value = ACCOUNT_PATH, method = RequestMethod.GET)
+    public ResponseEntity<?> getListUser(){
+        boolean isAdmin = accountService.adminCheck();
+        if(!isAdmin){
+            return ResponseEntity.status(403).body("Ban chưa được cấp quyền...");
+        }
+        return ResponseEntity.ok(accountManagerService.findAll());
+    }
+
     @RequestMapping(value = ACCOUNT_PATH_WITH_ID, produces = "application/json", consumes = "application/json", method = RequestMethod.PUT)
     public ResponseEntity<?> update(@RequestBody Account account, @PathVariable Long id) {
         boolean isAdmin = accountService.adminCheck();
         if(!isAdmin){
-            return ResponseEntity.status(403).body("....");
+            return ResponseEntity.status(403).body("Ban chưa được cấp quyền...");
         }
         return ResponseEntity.ok(accountManagerService.update(account, id));
     }
 
     @RequestMapping(value = ACCOUNT_PATH_WITH_ID, method = RequestMethod.DELETE)
     public ResponseEntity<?> delete(@PathVariable Long id){
+        boolean isAdmin = accountService.adminCheck();
+        if(!isAdmin){
+            return ResponseEntity.status(403).body("Ban chưa được cấp quyền...");
+        }
         accountManagerService.delete(id);
         return ResponseEntity.ok("Deleted");
     }
-    @RequestMapping(value = ACCOUNT_PATH, method = RequestMethod.GET)
-    public Page<Account> findPage(@RequestParam int page, @RequestParam int size) {
-        PageRequest pageRequest = PageRequest.of(page, size);
-        return repository.findAll(pageRequest);
-    }
+//    @RequestMapping(value = ACCOUNT_PATH, method = RequestMethod.GET)
+//    public Page<Account> findPage(@RequestParam int page, @RequestParam int size) {
+//        PageRequest pageRequest = PageRequest.of(page, size);
+//        return repository.findAll(pageRequest);
+//    }
 }
